Move todo handoff gate into control plane
