+++ loopx/canary/planner.py
@@ -318,7 +318,7 @@ CURRENT_REPO_PROFILES: tuple[dict[str, Any], ...] = (
             "successor",
             "loopx/control_plane/scheduler/monitor_todo.py",
             "loopx/control_plane/scheduler/scheduler_hint.py",
-            "loopx/todo_handoff_gate.py",
+            "loopx/control_plane/todos/handoff_gate.py",
         ),
         "checks": [
             {
